feat: continuity statements for the bundle of continuous linear maps (#25383)
The continuity statements I am adding are counterparts of already existing ones for smooth maps.

Also solve a TODO by dropping the type synonym `Bundle.ContinuousLinearMap`. It was useful a long time ago when the topology on the fibers was built using the norm on the model space. This is not the case since we've endowed all spaces of continuous linear maps on topological vector spaces with the topology of uniform convergence on bounded sets.