chore(Tactic/Linarith): put linarith declarations into `Mathlib.Tactic` namespace (#25955)
The `Name.isMetaProgramming` function is used to exclude lemmas from library search tactics. Thus, it is important that lemmas that are only meant for use in a tactic implementation have a name flagged by `Name.isMetaProgramming`. So, this PR renames everying in `Mathlib/Tactic/Linarith` to the namespace `Mathlib.Tactic.Linarith` (which previously was just `Linarith`)