chore: rename file (#26135)
In preparation for an upcoming PR, rename "Analysis/SpecialFunctions/Integrals.lean" to "Analysis/SpecialFunctions/Integrals/Basic.lean." Rationale: I plan to upload code that computes the integral of `log(sin z)`. The code should probably not be included in "Analysis/SpecialFunctions/Integrals.lean" because the methods are different, and my code requires massive new imports. In this setting, I feel it makes sense to open a new directory for code where integrals are computed.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.