feat(CategoryTheory/Monoidal): Day convolution of functors (#25755)
Given monoidal categories `C`,`V` and functors `F G : C ⥤ V`, we introduce a typeclass `[DayConvolution F G]` that bundles the data of a functor `F ⊛ G`, as well as the data required to exhibit `F ⊛ G` as a pointwise left Kan extension of `F ⊠ G` (the external product of `F` and `G`) along the tensor product of `C`.

We give basic API  around this class.

We also define a class `DayConvolutionUnit U` which contains the data that `U` is pointwise left Kan extended from `fromPUnit (𝟙_ V)` along `fromPUnit (𝟙_ C)`. Again, we give basic API around this class.

Properties like construction of associators and unitors isomorphisms (and the pentagon/triangle identities) will be the subject of subsequent PRs.