feat: integral of log ∘ sin (#26172)
Compute special values of the integral of `log ∘ sin`. Given that the indefinite integral involves the dilogarithm, this can be seen as computing special values of `Li₂`.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.