refactor: restate `(T35|T6)Space` using `T0Space` (#26065)
Currently, T3.5 and T6 spaces are defined as completely regular and perfectly normal T1 spaces, respectively. However, in both cases, the assumption of T1 can be weakened to T0.

Note: T4 and T5 spaces are also defined as normal and completely normal T1 spaces, respectively, but in these cases, the T1 assumption cannot be weakened to T0. This is because the Sierpiński space (`Prop`) is a completely regular T0 space, but it is not an R0 space, and therefore not a T1 space.



Co-authored-by: Miyahara Kō <52843868+Komyyy@users.noreply.github.com>