feat(Geometry/Euclidean/Incenter): touchpoints and tangency (#25935)
Define touchpoints for an exsphere / insphere and set up basic lemmas about exspheres / inspheres being tangent to the faces at the touchpoints (previously only present in a form talking about distance, not in a form directly using definitions about tangency). There's more to be done in followups, in particular showing the signs of the barycentric coordinates of touchpoints (so that e.g. the touchpoints for the insphere are in the interiors of the faces), but I think setting up the definition here is a reasonable start.