chore: use `have` instead of `let` for some propositions (#26493)
This PR uses `have` instead of `let` in a number of places, found during the development of https://github.com/leanprover/lean4/pull/8373

Adds one adaptation note. The unused have/suffices linter is interacting poorly with Qq; the `have` is unused in the underlying term, but it's used during elaboration.