feat: make the `labels_from_comment` action more extensible (#26456)
The revised version of this script manages the labels `awaiting-author`, `WIP` and `easy` from comments.  If a PR comment contains of one of the three mentioned labels, optionally preceded by a single `-` *as a complete line*, then the action will add/remove the corresponding label.  If multiple lines match the same label, only the latest mention will be used.

The line may contain whitespace, besides the sign and the label, but nothing else.

For instance, if the comment is
```
A
 awaiting-author
-easy
some WIP text
       easy  
-awaiting-author
```
then the script would
* *add* the `easy` label;
* *remove* the `awaiting-author` label;
* do nothing with the `WIP` label, since it does not appear alone on any line.

Note: the previous action would automatically remove all remaining labels, since the old set was mutually exclusive.  The expectation is that the current set of labels will be increased, so that mutual exclusivity is no longer expected.