fix(Cache): check for urls in the remote before querying for the url  (#26551)
It can happen, as with `gh pr checkout`, that the remote is not a regular name but a URL. Currently, `getRepoFromRemote` calls `git remote get-url` which chokes on an actual URL breaking the logic for getting the repo name.

Here we provide a basic spot check for whether the remote is already a URL. If so, we return it directly. Otherwise, we continue with the existing logic.

Note: we comment out any actually changes to the logic but include a print statement to show what the logic would have been. We do this out of an abundance of caution, as we don't want to break existing CI. This needs to be tested sufficiently and then the commented code can be uncommented.



Co-authored-by: Matthew Robert Ballard <100034030+mattrobball@users.noreply.github.com>