feat(CategoryTheory/Limits/Shapes/Pullback): categorical pullbacks (#26366)
Given a functor `F : A ⥤ B` and `G : C ⥤ B`, `CategoricalPullback F G` is the category of triples `(a : A, c : C, e : F.obj a ≅ G.obj b)`. This construction is to be thought of as an explicit (unbundled) model of the pseudolimit of the cospan in `Cat` defined by `F` and `G`. To justify this claim, after giving some basic API on `CategoricalPullback`, we construct an equivalence of categories between `X ⥤ CategoricalPullback F G` and `CategoricalPullback (whiskeringRight X A B|>.obj F) (whiskeringRight X C B|>.obj G)`, the latter is by definition the categories of triples `(U : X ⥤ A`, `V : X ⥤ C`, `U ⋙ F ≅ V ⋙ G)`, i.e the category of "categorical commutative squares" with top left corner `X` and lying over the cospan defined by `F` and `G`. This equivalence realizes the universal property of the categorical pullback.

This PR is first of a series that aims at bringing API for (2,1)-categorical pullbacks of categories. Further work on the subject includes
* Definition of a type-class extending `CatCommSq _ _ F G` that bundles the assumption that a categorical commutative square is a categorical pullback square, by bundling the data of an equivalence with the categorical pullback of `F` and `G` compatible with the 2-isomorphisms.
* Full 2-functoriality and equivalence-invariance of this notion.
* Horizontal and vertical pasting calculus for categorical pullback squares.