feat(Analysis.Normed.Unbundled.SpectralNormUnique): prove unique extension theorem (#26537)
Let `K` be a field complete with respect to a nontrivial nonarchimedean multiplicative norm and `L/K` be an algebraic extension. We show that the spectral norm on `L` is a nonarchimedean multiplicative norm, and any power-multiplicative `K`-algebra norm on `L` coincides with the spectral norm. More over, if `L/K` is finite, then `L` is a complete space.
This result is [S. Bosch, U. Güntzer, R. Remmert,*Non-Archimedean Analysis* (Theorem 3.2.4/2)][bosch-guntzer-remmert].