feat(CategoryTheory/Monoidal/ExternalProduct): external product of pointwise left Kan extensions of functors (#25754)
Given a functor `H' : D' ⥤ V` to a monoidal category that is a pointwise left Kan along `L : D ⥤ D'`, we show that for a functor `K : E ⥤ V` and under suitable preservations of colimits by the tensor product of `V`, the functor `H' ⊠ K : D' × E ⥤ V` is a pointwise left Kan extension along `L.prod (𝟭 E)`. We show a similar statement for the functor `K ⊠ H'`.

This property is crucial for proving properties of Day convolutions of functors from a monoidal category to `V`.