refactor(CategoryTheory/Monoidal/Mod_): refactor `Mod_` using `MonoidalLeftAction` (#25762)
Generalize the definition of `CategoryTheory.Monoidal.Mod_Class` and `CategoryTheory.Monoidal.Mod_` to modules in `D` over a monoid object in `C`, where `C` acts monoidally on `D`. 

At the same time, we refactor `Mod_` in a way similar to the one `Mon_` has been refactored in #24646. i.e we favor `Mod_Class` in the definition and `Mod_`, and introduce a Prop-class `IsMod_Hom` to define morphisms of modules.