feat: restriction of scalars for iterated Fréchet derivatives (#26353)
Establish standard theorems on restriction of scalars for iterated Fréchet derivatives, comparing iterated derivatives with respect to a field `𝕜'` to iterated derivatives with respect to a subfield `𝕜 ⊆ 𝕜'`. The results are analogous to those found in
`Mathlib.Analysis.Calculus.FDeriv.RestrictScalars`.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane.