feat(CategoryTheory/Monoidal/DayConvolution): left and right unitors for Day convolution (#25757)
Given a `DayConvolutionUnit U` a functor `F : C ⥤ V` and a Day convolution of `U` and `F`, we construct the left unitor isomorphism. `U ⊛ F ≅ F`. Similarly, given a Day convolution of `F` and `U`, we construct the right unitor.

We characterize the isomorphisms, show their naturality, and show that the triangle identity is satisfied.