feat: implement harmonic functions (#26721)
Define harmonic functions on real, finite-dimensional, inner product spaces.

This material is used in [Project VD](https://github.com/kebekus/ProjectVD), which aims to formalize Value Distribution Theory for meromorphic functions on the complex plane. It is one of the ingredients in the proof of Jensen's Formula in complex analysis.