feat(Algebra/Homology): truncations of ConnectData.cochainComplex (#26520)
If `h : ConnectData K L` with `K` a chain complex and `L` a cochain complex, the "restriction" of the cochain complex `h.cochainComplex` in nonnegative degrees identifies to `L` and the restriction in negative degrees identifies to `K`. This allows to compute the homology of `h.cochainComplex` in degrees different from `0` and `-1`.