feat(CategoryTheory): add dinatural transformations (continuedx2) (#26854)
This is a refactor of https://github.com/leanprover-community/mathlib4/pull/26577, which itself was a refactor of https://github.com/leanprover-community/mathlib4/pull/8118, with permission by @iwilare. It contains the definition of a dinatural transformation. There is no current future plans for these but maybe I or @iwilare will use these in the future.

I did very little, credits go to the other two authors.

Co-authored-by: [Andrea Laretto](https://github.com/iwilare) [iwilare@gmail.com](mailto:iwilare@gmail.com)
Co-authored-by:  [Brendan Seamas Murphy](https://github.com/Shamrock-Frost) [bsmurphy@math.utah.edu](mailto:bsmurphy@math.utah.edu)