feat(CategoryTheory/Functor/KanExtension) : Kan extension along equivalences of categories (#26906)
In this PR, we record as instances the somewhat trivial fact that if `L : C ≌ D` is an equivalence of categories and `F₀ ≅ L.functor ⋙ F₁` any natural isomorphism where `F₀` is a functor out of `C` and `F₁` a functor out of `D`, then the corresponding extension is a left Kan extension. The dual fact for right Kan extensions is recorded.