chore(Order/IsNormal): change default constructor (#26903)
The default constructor of `Order.IsNormal` currently contains a redundant hypothesis. The idea was that it might often be easier to reason in terms of `IsLUB` rather than unfolding everything down to inequalities. My work on #26900 suggests however that's not the case.

I've also added the theorem `isNormal_iff`, which takes care of unfolding the definition. (Note that it matches that given by `Ordinal.isNormal_iff_strictMono_limit`!)