feat: custom elaborators for differential geometry
Add two sets of custom elaborators for differential geometry, to
- infer the model with corners in `MDifferentiable{WithinAt,At,On}` or `ContMDiff{WithinAt,At,On}`
  expressions from the local context, and provide shorter notation in that setting
- convert a fibre bundle section from a dependent section to a function into the total space,
  making working with differentiability of sections of fibre bundles less cumbersome

This was [discussed on zulip](...), to overall broad consensus.

This code has substantial tests and has been tried in the branch in #26221 (uncovering minor issues,
but not substantial ones). The implementation itself is of prototype quality and will be cleaned up
in the future. Once it is more robust, more differential geometry files can be converted to use it.

-----------

I agree that merging not fully polished code is not ideal, but propose doing so for two reasons:
- the work in #26221 used these elaborators extensively (to test them).
  Reverting those changes, merging the code, only to re-apply them is not a good use of time.
  Merging this PR now enables merging the work in #26221 soon.
- the feature is clearly useful and not going away: in the files that use it, it works well already.
  Holding off on further adopting, but merging it now seems like a good compromise.