feat: solution to ODE is Lipschitz with respect to the initial condition (#26392)
This PR continues the work from #20733.

Original PR: https://github.com/leanprover-community/mathlib4/pull/20733

I formalise a corollary to the Picard-Lindelöf theorem: the local solution to an ODE is Lipschitz continuous with respect to the initial condition. This in particular means that the local flow to the associated vector field is continuous.