feat: length of a path in a manifold (#26778)
Consider a manifold in which the tangent spaces have an enormed structure. Then one defines
`pathELength γ a b` as the length of the path `γ : ℝ → M` between `a` and `b`, i.e., the integral
of the norm of its manifold derivative. We develop a reasonable API around this notion

We also define `riemannianEDist x y` as the infimum of the length of paths from `x` to `y`. We show that it is symmetric and satisfies the triangle inequality.