fix(RefinedDiscrTree): remove funny `transparency` handling (#26531)
This PR fixes some funny behaviour of `RefinedDiscrTree` discovered at #26484.

This PR aligns the behaviour with Lean core's `DiscrTree`. The original reason for this special treatment was to avoid a mathlib breakage. But now it turns out that this special trearment also causes a mathlib breakage (but this one is harder to fix).

The breakage that was the original reason for this behaviour is easily fixed using a `dsimp`.

#26484 depends on this PR