feat: improve quality of proofs generated by `congr(...)`, add caching, add support for overapplied functions (#25851)
This PR makes the following improvements to the `congr(...)` term elaborator:
- It now uses `congr`/`congrFun` as much as possible instead of creating giant HEq-based congruence proofs.
- It now looks for common prefixes for applications, to skip unnecessary work proving that these parts of an application are equal.
- Intermediate results are cached.
- Overapplied functions are now supported.

An overapplied function is a function with more arguments than the arity of the function itself. For example, in `id id 2` the first `id is overapplied. The return type is conditionally a pi type depending on the values of the arguments.

Fixes Jz Pan's issue reported [on Zulip](https://leanprover.zulipchat.com/#narrow/channel/287929-mathlib4/topic/Strange.20error.20on.20.60congr.60.20macro/near/523998660).

Modifies a couple uses of `congr(...)` in `SimplicialSet.NerveAdjunction` to take more advantage of it.