feat(combinatorics): Add weight function for Quiver.Path
This PR introduces a generic weight function for paths in a quiver, where edge weights are elements of a monoid. This is the first in a series of planned contributions to Quiver.Path API, with the ultimate goal of PRing the formalized Perron-Frobenius theorem for primitive and irreducible matrices here .

https://leanprover.zulipchat.com/#narrow/channel/116395-maths/topic/Formalizing.20Perron-Frobenius/with/525516636

The Perron-Frobenius theorem relies on analyzing the powers of a matrix, which correspond to the number (or total weight) of paths of a given length in the graph represented by the matrix. A robust API for path manipulation is therefore a prerequisite.

This API is built on Quiver because it is fundamentally concerned with the properties of paths as sequences of specific, identifiable arrows (a ⟶ b). The Digraph structure, which models adjacency as a relation, abstracts away the identity of individual edges and would not support the construction and manipulation of paths in the required manner (not in my attempts at least). Similarly, Graph is designed for undirected graphs. Therefore, I think Quiver is the appropriate foundational layer for this work

Future Work: Subsequent PRs will (partly) build on this foundation by introducing:

- Definitions for the vertices of a path (Path.vertices) and core properties.
- Lemmas for path decomposition and splitting.
- A formalization of simple paths and cycles, including cycle extraction.
- The theorem that a shortest path is always simple.